requirements.txt
# ShadowDance - LangSmith observability for Unitree robot SDKs

# Core dependencies
langsmith>=0.1.0
python-dotenv>=1.0.0
httpx>=0.25.0

# Optional: OpenAI client (for LLM tracing)
openai>=1.0.0

# Development dependencies (optional)
# pytest>=7.0.0
# ruff>=0.1.0
# mypy>=1.0.0
